renamed TextParserToFrequencyDictionary -> TextAnalyzer
