+++ TagsCloudContainer/App/TextAnalyzer/SimpleTextParser.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
-using TagsCloudContainer.Infrastructure.TextParserToFrequencyDictionary;
+using TagsCloudContainer.Infrastructure.TextAnalyzer;
 
-namespace TagsCloudContainer.App.TextParserToFrequencyDictionary
+namespace TagsCloudContainer.App.TextAnalyzer
 {
     internal class SimpleTextParser : ITextParser
     {
